Move NamedayCalendar specific classes to respective package
