+++ mobile/src/main/java/com/alexstyl/specialdates/datedetails/DateDetailsAdapter.java
@@ -14,8 +14,8 @@ import com.alexstyl.specialdates.events.DayDate;
 import com.alexstyl.specialdates.events.bankholidays.BankHoliday;
 import com.alexstyl.specialdates.events.bankholidays.BankholidayCalendar;
 import com.alexstyl.specialdates.events.bankholidays.BankHolidaysPreferences;
-import com.alexstyl.specialdates.events.namedays.NamedayCalendar;
-import com.alexstyl.specialdates.events.namedays.NamedayCalendarProvider;
+import com.alexstyl.specialdates.events.namedays.calendar.NamedayCalendar;
+import com.alexstyl.specialdates.events.namedays.calendar.NamedayCalendarProvider;
 import com.alexstyl.specialdates.events.namedays.NamedayLocale;
 import com.alexstyl.specialdates.events.namedays.NamedayPreferences;
 import com.alexstyl.specialdates.events.namedays.NamesInADate;
