+++ mobile/src/main/java/com/alexstyl/specialdates/events/bankholidays/BankHolidayRepository.java
@@ -4,7 +4,7 @@ import android.support.annotation.Nullable;
 
 import com.alexstyl.specialdates.date.Date;
 import com.alexstyl.specialdates.events.DayDate;
-import com.alexstyl.specialdates.events.namedays.EasterCalculator;
+import com.alexstyl.specialdates.events.namedays.calendar.EasterCalculator;
 
 import java.util.List;
