+++ mobile/src/main/java/com/alexstyl/specialdates/events/namedays/NamedayBundle.java
@@ -1,6 +1,7 @@
 package com.alexstyl.specialdates.events.namedays;
 
 import com.alexstyl.specialdates.events.DayDate;
+import com.alexstyl.specialdates.events.namedays.calendar.Node;
 
 import java.util.ArrayList;
