+++ mobile/src/main/java/com/alexstyl/specialdates/events/namedays/calendar/CharacterNode.java
@@ -1,7 +1,9 @@
-package com.alexstyl.specialdates.events.namedays;
+package com.alexstyl.specialdates.events.namedays.calendar;
 
 import com.alexstyl.gsc.Index;
 import com.alexstyl.specialdates.events.DayDate;
+import com.alexstyl.specialdates.events.namedays.Dates;
+import com.alexstyl.specialdates.events.namedays.NameCelebrations;
 
 import java.text.Collator;
 import java.util.ArrayList;
