+++ mobile/src/main/java/com/alexstyl/specialdates/events/namedays/calendar/GreekNamedays.java
@@ -1,6 +1,9 @@
-package com.alexstyl.specialdates.events.namedays;
+package com.alexstyl.specialdates.events.namedays.calendar;
 
 import com.alexstyl.specialdates.events.DayDate;
+import com.alexstyl.specialdates.events.namedays.NameCelebrations;
+import com.alexstyl.specialdates.events.namedays.NamedayBundle;
+import com.alexstyl.specialdates.events.namedays.NamesInADate;
 import com.novoda.notils.exception.DeveloperError;
 
 import java.util.ArrayList;
