+++ mobile/src/main/java/com/alexstyl/specialdates/events/namedays/calendar/NamedayCalendar.java
@@ -1,6 +1,12 @@
-package com.alexstyl.specialdates.events.namedays;
+package com.alexstyl.specialdates.events.namedays.calendar;
 
 import com.alexstyl.specialdates.events.DayDate;
+import com.alexstyl.specialdates.events.namedays.Dates;
+import com.alexstyl.specialdates.events.namedays.NameCelebrations;
+import com.alexstyl.specialdates.events.namedays.NamedayBundle;
+import com.alexstyl.specialdates.events.namedays.NamedayLocale;
+import com.alexstyl.specialdates.events.namedays.NamesInADate;
+import com.alexstyl.specialdates.events.namedays.SpecialNamedaysStrategy;
 
 import java.util.ArrayList;
 import java.util.List;
