+++ mobile/src/main/java/com/alexstyl/specialdates/events/namedays/calendar/NamedayCalendarProvider.java
@@ -1,6 +1,11 @@
-package com.alexstyl.specialdates.events.namedays;
+package com.alexstyl.specialdates.events.namedays.calendar;
 
 import android.content.Context;
+
+import com.alexstyl.specialdates.events.namedays.NamedayBundle;
+import com.alexstyl.specialdates.events.namedays.NamedayLocale;
+import com.alexstyl.specialdates.events.namedays.NamedayPreferences;
+import com.alexstyl.specialdates.events.namedays.SpecialNamedaysStrategy;
 
 import org.json.JSONArray;
