+++ mobile/src/main/java/com/alexstyl/specialdates/events/namedays/calendar/NoSpecialNamedaysStrategy.java
@@ -1,6 +1,9 @@
-package com.alexstyl.specialdates.events.namedays;
+package com.alexstyl.specialdates.events.namedays.calendar;
 
 import com.alexstyl.specialdates.events.DayDate;
+import com.alexstyl.specialdates.events.namedays.NameCelebrations;
+import com.alexstyl.specialdates.events.namedays.NamesInADate;
+import com.alexstyl.specialdates.events.namedays.SpecialNamedaysStrategy;
 
 import java.util.ArrayList;
