+++ mobile/src/main/java/com/alexstyl/specialdates/events/namedays/calendar/Node.java
@@ -1,6 +1,7 @@
-package com.alexstyl.specialdates.events.namedays;
+package com.alexstyl.specialdates.events.namedays.calendar;
 
 import com.alexstyl.specialdates.events.DayDate;
+import com.alexstyl.specialdates.events.namedays.NameCelebrations;
 
 public interface Node {
