+++ mobile/src/main/java/com/alexstyl/specialdates/events/namedays/calendar/RomanianEasterSpecialCalculator.java
@@ -1,4 +1,4 @@
-package com.alexstyl.specialdates.events.namedays;
+package com.alexstyl.specialdates.events.namedays.calendar;
 
 import com.alexstyl.specialdates.events.DayDate;
