+++ mobile/src/main/java/com/alexstyl/specialdates/events/namedays/calendar/SpecialNamedaysHandlerFactory.java
@@ -1,4 +1,7 @@
-package com.alexstyl.specialdates.events.namedays;
+package com.alexstyl.specialdates.events.namedays.calendar;
+
+import com.alexstyl.specialdates.events.namedays.NamedayLocale;
+import com.alexstyl.specialdates.events.namedays.SpecialNamedaysStrategy;
 
 import org.json.JSONArray;
