+++ mobile/src/test/java/com/alexstyl/specialdates/events/namedays/RomanianEasterSpecialCalculatorTest.java
@@ -1,6 +1,7 @@
 package com.alexstyl.specialdates.events.namedays;
 
 import com.alexstyl.specialdates.events.DayDate;
+import com.alexstyl.specialdates.events.namedays.calendar.RomanianEasterSpecialCalculator;
 
 import java.util.ArrayList;
 import java.util.List;
