+++ mobile/src/test/java/com/alexstyl/specialdates/namedays/CharacterNodeTest.java
@@ -1,7 +1,7 @@
 package com.alexstyl.specialdates.namedays;
 
-import com.alexstyl.specialdates.events.namedays.CharacterNode;
-import com.alexstyl.specialdates.events.namedays.Node;
+import com.alexstyl.specialdates.events.namedays.calendar.CharacterNode;
+import com.alexstyl.specialdates.events.namedays.calendar.Node;
 
 public class CharacterNodeTest extends NodeTest {
