+++ mobile/src/test/java/com/alexstyl/specialdates/namedays/SoundNodeTest.java
@@ -1,7 +1,7 @@
 package com.alexstyl.specialdates.namedays;
 
-import com.alexstyl.specialdates.events.namedays.Node;
-import com.alexstyl.specialdates.events.namedays.SoundNode;
+import com.alexstyl.specialdates.events.namedays.calendar.Node;
+import com.alexstyl.specialdates.events.namedays.calendar.SoundNode;
 
 public class SoundNodeTest extends NodeTest {
